Update and modularize file structure to base DebOps Contrib docs on it
