+++ docs/_bin/gen_ansible_modules_rst_defs
