+++ docs/_bin/gen_rst_defs_using_debops_api
