+++ docs/_lib/sphinxcontrib/debops/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+
+from .func import *      # pylint: disable=wildcard-import
