+++ docs/conf.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 #
 # DebOps documentation build configuration file
-# Copyright (C) 2014-2016 DebOps Project http://debops.org/
+# Copyright (C) 2014-2016 DebOps Project https://debops.org/
 #
 # This file is execfile()d with the current directory set to its
 # containing dir.
@@ -12,150 +12,21 @@
 # All configuration values have a default; values that are commented out
 # serve to show the default.
 
-import sys
-import os
-import fnmatch
-import re
-
-#  import git
-
-# Generate documentation on the fly based on Ansible default variables
-import yaml2rst
-
-for element in os.listdir('ansible/roles'):
-    if os.path.isdir('ansible/roles/' + element):
-        yaml2rst.convert_file(
-            'ansible/roles/' + element + '/defaults/main.yml',
-            'ansible/roles/' + element + '/docs/defaults.rst',
-            strip_regex=r'\s*(:?\[{3}|\]{3})\d?$',
-            yaml_strip_regex=r'^\s{66,67}#\s\]{3}\d?$',
-        )
-
-from subprocess import call, check_output
-call(['bin/sphinx_conf_pre_hook'])
-
-on_rtd = os.environ.get('READTHEDOCS', None) == 'True'
-
-# The theme to use for HTML and HTML Help pages.  See the documentation for
-# a list of builtin themes.
-html_theme = 'default'
-
-if not on_rtd:  # only import and set the theme if we're building docs locally
-    import sphinx_rtd_theme
-    html_theme = 'sphinx_rtd_theme'
-    html_theme_path = [sphinx_rtd_theme.get_html_theme_path()]
-
-# Fix "Edit on GitHub" links (((
-# Jinja2 Support is only basic Jinja2 without all the good stuff from Ansible. So I am not gonna mess with that or try to extend it as in:
-# https://stackoverflow.com/questions/36019670/removing-the-edit-on-github-link-when-using-read-the-docs-sphinx-with-readthed
-# What I am gonna do instead is just recompute source file to URL map in Python and job done.
-#
-# git_repo.iter_submodules() fails with "unknown encoding: -----BEGIN PGP SIGNATURE-----"
-
-
-def find_files(directory, pattern):
-    for root, dirs, files in os.walk(directory):
-        for basename in files:
-            if fnmatch.fnmatch(basename, pattern):
-                filename = os.path.join(root, basename)
-                yield filename
-
-
-def get_source_file_to_url_map(start_dir='.'):
-    skip_patterns = [
-        r'debops-keyring/docs/entities(?:\.rst)$',  # Auto generated.
-        r'ansible/roles/debops[^/]+$',  # Legacy.
-    ]
-
-    source_file_to_url_map = {}
-    repo_dir_to_url_map = {}
-    list_of_submod_paths = []
-
-    cur_dir = os.path.abspath(os.path.dirname(__file__))
-
-    for submodule_path in check_output(['git', 'submodule', '--quiet', 'foreach', 'pwd']).split('\n'):
-        if submodule_path.startswith(cur_dir):
-            submodule_path = submodule_path[len(cur_dir):].lstrip('/')
-        list_of_submod_paths.append(submodule_path)
-
-    for source_file_name in find_files('.', '*.rst'):
-        pagename_source_file = source_file_name.lstrip('/.')
-
-        skip = False
-        for skip_pattern in skip_patterns:
-            if re.search(skip_pattern, pagename_source_file):
-                #  print(pagename_source_file)
-                skip = True
-                break
-
-        if skip:
-            continue
-
-        dir_path = os.path.dirname(source_file_name)
-        if len(dir_path) > 2:
-            dir_path = dir_path.lstrip('/.')
-
-        # Can also contain subdirs in a repo but this optimization should already
-        # get factor 10 in performance for git Invokation.
-        if dir_path not in repo_dir_to_url_map:
-            #  git_repo = git.Repo(dir_path)
-            #  repo_dir_to_url_map[dir_path] = git_repo.remotes.origin.url
-            for remote_line in check_output(['git', '-C', dir_path, 'remote', '-v']).split('\n'):
-                remote_item = re.split(r'\s', remote_line)
-                if remote_item[0] == 'origin' and remote_item[2] == '(fetch)':
-                    base_url = remote_item[1]
-                    if base_url.endswith('.git'):
-                        base_url = base_url[:-4]
-                    repo_dir_to_url_map[dir_path] = base_url
-                    #  print(repo_dir_to_url_map[dir_path])
-
-        relative_pagename = pagename_source_file
-
-        if relative_pagename in ['index.rst', 'ansible/roles/index.rst']:
-            relative_pagename = 'docs/' + relative_pagename
-
-        for submod_path in list_of_submod_paths:
-            if pagename_source_file.startswith(submod_path + '/'):
-                relative_pagename = pagename_source_file[len(submod_path):].lstrip('/')
-
-        # Does not work for legacy roles yet. Disabled.
-        #  if re.match(r'docs/copyright(?:\.rst)$', relative_pagename, flags=re.I):
-        #      relative_pagename = 'COPYRIGHT'
-
-        if re.match(r'docs/readme(?:\.rst)$', relative_pagename, flags=re.I):
-            relative_pagename = 'README.rst'
-
-        if re.match(r'docs/changelog(?:\.rst)$', relative_pagename, flags=re.I):
-            relative_pagename = 'CHANGES.rst'
-
-        pagename_source_file = re.sub(r'\.rst$', '', pagename_source_file)
-        source_file_to_url_map[pagename_source_file] = {
-            'url': repo_dir_to_url_map[dir_path],
-            'pagename': relative_pagename,
-        }
-        #  print('{}: {}'.format(pagename_source_file, source_file_to_url_map[pagename_source_file]))
-
-
-    #  print(source_file_to_url_map)
-    #  import pprint
-    #  pprint.pprint(source_file_to_url_map)
-    return source_file_to_url_map
-
-html_context = {
-    'display_github': True,  # Add 'Edit on Github' link instead of 'View page source'
-    'last_updated': True,
-    'commit': False,
-    'source_file_to_url_map': get_source_file_to_url_map()
-}
-
-# https://stackoverflow.com/a/21909382
-#  import sphinx.application.TemplateBridge
-# )))
-
 # If extensions (or modules to document with autodoc) are in another directory,
 # add these directories to sys.path here. If the directory is relative to the
 # documentation root, use os.path.abspath to make it absolute, like shown here.
+#
+import os
+import sys
 # sys.path.insert(0, os.path.abspath('.'))
+
+sys.path.insert(0, os.path.abspath('_lib'))
+from sphinxcontrib import debops
+
+debops.yaml2rst_role_defaults('ansible/roles/')
+
+from subprocess import call
+call(['_bin/sphinx_conf_pre_hook'])
 
 # -- General configuration ------------------------------------------------
 
@@ -232,8 +103,8 @@ pygments_style = 'sphinx'
 # for more details.
 highlight_language = 'YAML'
 
-## TODO: Change later to this when it can handle:
-## enabled: '{{ True if (owncloud_database_name != owncloud_database_user) else False }}'
+# TODO: Change later to this when it can handle:
+# enabled: '{{ True if (owncloud_database_name != owncloud_database_user) else False }}'
 #  highlight_language = 'YAML+Jinja'
 
 # A list of ignored prefixes for module index sorting.
@@ -245,6 +116,29 @@ highlight_language = 'YAML'
 
 # -- Options for HTML output ----------------------------------------------
 
+# The theme to use for HTML and HTML Help pages.  See the documentation for
+# a list of builtin themes.
+html_theme = 'default'
+
+on_rtd = os.environ.get('READTHEDOCS', None) == 'True'
+
+if not on_rtd:  # only import and set the theme if we're building docs locally
+    import sphinx_rtd_theme
+    html_theme = 'sphinx_rtd_theme'
+    html_theme_path = [sphinx_rtd_theme.get_html_theme_path()]
+
+html_context = {
+    'display_github': True,  # Add 'Edit on Github' link instead of 'View page source'
+    'last_updated': True,
+    'commit': False,
+    'source_file_to_url_map': debops.get_source_file_to_url_map(
+        skip_patterns=[
+            r'debops-keyring/docs/entities(?:\.rst)$',  # Auto generated.
+            r'ansible/roles/debops[^/]+$',  # Legacy.
+        ],
+    )
+}
+
 # Theme options are theme-specific and customize the look and feel of a theme
 # further.  For a list of options available for each theme, see the
 # documentation.
@@ -255,7 +149,7 @@ highlight_language = 'YAML'
 
 # The name for this set of Sphinx documents.  If None, it defaults to
 # "<project> v<release> documentation".
-#html_title = None
+# html_title = None
 
 # A shorter title for the navigation bar.  Default is the same as html_title.
 #html_short_title = None
